cart item deliver address ,order item,order ,wilayah

+++ app/tag/controller.js
@@ -2,7 +2,14 @@ const Tag = require('./model');
 
 async function store(req, res, next){
     try {
-        //dapatkan data dari reques yang dikirimkan client
+       
+        let policy = policyFor(req.user);
+        if(!policy.can('create','Tag')){
+         return res.json({
+             error:1,
+             message:`Anda tidak memiliki akses untuk membuat Tag`
+         });
+        } //dapatkan data dari reques yang dikirimkan client
         let payload = req.body;
         //buat objek tag baru berdasarkan payload
         let tag = new Tag(payload)
@@ -25,6 +32,13 @@ async function store(req, res, next){
 
 async function update(req, res, next){
     try {
+        let policy = policyFor(req.user);
+        if(!policy.can('update','Tag')){
+         return res.json({
+             error:1,
+             message:`Anda tidak memiliki akses untuk mengupdate Tag`
+         });
+        }
         //dapatkan data dari reques yang dikirimkan client
         let payload = req.body;
         //buat objek tag baru berdasarkan payload
@@ -47,7 +61,15 @@ async function update(req, res, next){
 
 async function destroy(req, res, next){
     try {
+        let policy = policyFor(req.user);
+        if(!policy.can('delete','Tag')){
+         return res.json({
+             error:1,
+             message:`Anda tidak memiliki akses untuk menghapus Tag`
+         });
+        }
         let tag = await Tag.findOneAndDelete({_id:req.params.id}) 
+        return res.json(tag);
     } catch (err) {
         next(err)
     }
